+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # Changelog
 
+## 2026-09-09 18:04 HKT
+
+- No differences between official feed and PDF baseline.
+
+## 2026-09-09 18:00 HKT
+
+- No differences between official feed and PDF baseline.
+
+## 2026-09-09 17:59 HKT
+
+- No differences between official feed and PDF baseline.
+
 ## 2026-09-09 16:53 HKT
 
 - No differences between official feed and PDF baseline.